bug[next][dace]: fix symbol mapping in lowering of let-statements (#2251)
This PR solves a bug which slipped through in #2240. The mapping of
symbols for array shape needs to follow the new canonical order of field
dimensions, for which the vertical dimensions follow the local one.
Therefore, we apply the gt4py `order_dimensions()` utility on the full
field domain, which includes also the local dimension.

---------

Co-authored-by: Hannes Vogt <vogt@hey.com>
Co-authored-by: Hannes Vogt <hannes@havogt.de>
Co-authored-by: Philip Mueller <philip.mueller@cscs.ch>