test[cartesian]: delete old/unused test utils (#2254)
## Description

This is a small follow-up from when we moved to the schedule tree in PR
https://github.com/GridTools/gt4py/pull/2067. Looks like we left some
test utils. `mypy` isn't complaining about the undefined import paths
because we skip it for all of tests and python isn't complaining because
the code never runs.

## Requirements

- [ ] All fixes and/or new features come with corresponding tests.
  N/A
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>