fix[cartesian]: Correct memlets for FrozenSDFG
When we freeze origin/domain into an SDFG, we tried to use python
trickery and be smart with `zip`. This led to hard to debug bugs showing
up in orchestrated UW. The specific case was a K-field with one data
dimension. That case would previously end up with a 3-dimensional range
memlet because `origin` and `shape` both were already 2d, including the
K dimension as well as the data dimension. On top of that, the data
dimension was added another time in the second loop about data
dimensions. Now, we cleanly separate cartesian and data dimenions,
skipping cartesian dimensions that aren't configured (as axis). As far
as I could see, `FieldInfo.boundary` always includes all 3 cartesian
dimesion, regardless if they are used or not.