fix[cartesian]: define domain symbols for all 3 dimensions
Always define all three cartesian dimension symbols. We'll need them in
the loops anyway. The previous implementation was relying on the fact
that we pass a 3d field into the stencils. However, this doesn't need to
be. For example, it is allowed to fill a 2d field with a given scalar.
In that particular case, there's no 3d field passed into the stencil and
thus we never defined the `__K` symbol.

Will be tested with the tests in the following commit.