feat[cartesian]: Add backend `dace:cpu_kfirst` (#2255)
## Description

In the process of performance optimizations, we realized that the naive
loop structure

```py
for k in range(K):   # vertical loop
    for i in range(I):  # horizontal loops
        for j in range(J):
            # do stencil computation here
```

of the `dace:cpu` backend

- might not be optimal in terms of cache access and
- isn't a 1:1 comparison with existing backends like `gt:cpu_kfirst`

We thus decided to write a small pass to flip vertical and horizontal
loops the loops conditionally. This is reflected in the
`dace:cpu_kfrist` backend, named after `gt:cpu_kfirst`.

Note: This is part of untangling the `milestone2` branch in PR
https://github.com/GridTools/gt4py/pull/2202.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
- [x] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>
Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>