refactor[dace]: Removed Dublicated Code (#2272)
Removed the `SingleStateGlobalSelfCopyElimination` from the `redundant_array_removers.py` file.

A class with the same name is also present in `single_state_global_self_copy_elimination.py`,
which is still present and was used in the end. Furthermore, the remaining implementation is
further developed.

This duplication was most likely introduced in [PR#2060](https://github.com/GridTools/gt4py/pull/2060].