fix[next][dace]: move cuda-codegen setting to gt4py config (#2271)
This PR is not fixing any known issue, since the setting set as default
in the modified in the dace integration branch (see
https://github.com/GridTools/dace/pull/6). However, for correctness, we
enforce this setting from gt4py, to be independent from the default
setting adopted by dace.

The setting for `match_exception` is already set in the gt4py
`auto_optimize` pipeline.