feat[next][dace]: extend CopyChainRemover to remove full-write copies (#2273)
The current design of `CopyChainRemover` allowed to remove temporary
arrays which are used as source node of full-read copies. Suppose a
temporary `__tmp` which is single-use data, with shape `(10,)`: `__tmp`
can be removed in `__tmp[0:10] -> A[x:x+10]`

With this change, `CopyChainRemover` can also be applied in the _write_
direction, that is to remove temporary arrays which are used as
destination node of full-write copies. For example, still using the
single-use temporary `__tmp` as above: `__tmp` can be removed in
`A[x:x+10] -> __tmp[0:10]`.