feat[next][dace]: collect sdfg execution time with tasklet (#2280)
This PR provides an alternative way to retrieve the SDFG execution time,
which corresponds to the stencil compute time metric in GT4Py.
The previous solution relied on SDFG instrumentation, which is a complex
framework designed to profile full-application SDFGs, to collect
different kinds of counters, and to report metrics across several
iterations of a benchmark run. This framework already exists in GT4Py,
we only need to timestamp the start and end time of the SDFG execution.
The solution presented in this PR is much more lightweight. More
importantly, it surpasses some limitation of the dace framework, namely
the small precision of the time measurements (dace truncates the SDFG
execution time to microseconds).
Another advantage is that the collection of compute time metric, and the
corresponding synchronization point for SDFG execution, can be enabled
or disabled by means of the GT4Py config option without recompiling the
program. This aligns the behavior of the dace backend with the gtfn
backend and makes it easier to collect metrics.