feat[dace][next]: Added Inline Fuser (#2283)
This PR adds an "Inline Fuser" functionality to the transformations.
This functionality is related to classical MapFusionVertical, however,
while in the classical case the two Maps have to be compatible, i.e. the
data generated by the first Map must be consumed in the _same_ iteration
by the second Map, the inline fuser is much more flexible, as it essentially
 removes the word _same_ from the definition. An extended description
can be found here: https://hackmd.io/ot702TuJT7OJXp07WveiPA