feat[next]: flexible metric collections (#2287)
This PR extends the `metrics` collection mechanism to track as different
metric sources the different compiled variants of a same program.

The main change in design is that metrics sources can contain any kind
of extra metadata and can collect metrics samples before they have a
properly assigned key for the global store of metrics sources. This
means that the key is not always defined in the outermost level of a
program callable, but in the innermost.

---------

Co-authored-by: Hannes Vogt <vogt@hey.com>