feat[dace][next]: Updated Splitting Fusion Transformations (#2296)
Essentially the `gt_horizontal_map_split_fusion()` and
`gt_vertical_map_split_fusion()` relied on the fact that when they are
called everything that _can_ be merged has been merged, thus they called
`MapFusionVertical` and `MapFusionHorizontal` directly without limiting
their scope. This worked well until the hooks, see
[PR#2294](https://github.com/GridTools/gt4py/pull/2294) where introduced
that allowed to limit the applications of `MapFusion`.

This PR modifies this my moving the MapFusion*` into the splitting
transformations themselves and limiting their applicability.
Furthermore, it introduces the possibility to pass a callback to inspect
if a split should be performed similar to the callbakcs `MapFusion*`
supports. Beside that this PR has the following impacts:
- Because fusion is now limited to what is split, not everything that
    should be fused is actually fused. This is not a problem in the current
    setup, as long as [PR#2284](https://github.com/GridTools/gt4py/pull/2284)
    is not merged.
- Du to how `MapFusion*` works it might be that even not everything that
    has been created by a split is fully merged.

The main reason for this PR is the update of the optimizer hooks, see
[PR#2294](https://github.com/GridTools/gt4py/pull/2294).