Avoid name conflict between tasklet connector and SDFG data
SDFG validation checks that there are no name conflicts between tasklet
connector and data entities in one SDFG. Because of some naming
convention, it could happen that `arg0` or `inp` were used as a tasklet
connector and as a data node, which would cause SDFG validation to fail.
This PR tries to avoid name conflicts by adding a prefix string to the
tasklet connectors.