fix[next][dace]: Avoid name conflict between tasklet connector and data node (#2306)
SDFG validation checks that there are no name conflicts between tasklet
connector and data entities in one SDFG. Because of some naming
convention, it could happen that `__arg0` was used as a tasklet
connector and as a data node, which would cause SDFG validation to fail.
This PR tries to avoid name conflicts by creating unique names for the
tasklet connectors.