feat[next][dace]: Enable customization of auto-optimize in dace backend workflow (#2295)
This PR exposes the `gt_auto_optimize()` parameters to application, in
order to allow setting up custom dace backends. The parameters include
regular function arguments as well as the new hook functions introduced
by #2293 and #2294.

Note that this PR removes the setting of `bindings.make_persistent` (and
just relies on its default value) because this option is going to be
deprecated soon in GT4Py.