feat[dace][next]: Demotion of Fields (#2288)
Adding the `demote_fields` argument to `gt_auto_optimizer()`.

The flag instructs the optimizer to handle SDFG arguments, which are
globals, as transients and by that allows more possible transformations.
After the has concluded they _might_ be put back, i.e. turned back into
a non-transient. However, it is not generally possible to put them back as
the optimization process might have modified them.

NOTE:
This feature is an experiential unsafe expert feature.

---------

Co-authored-by: edopao <edoardo16@gmail.com>