fix[next][dace]: Avoid name conflict between tasklet connector and data node (#2318)
This fix replaces the previous change in #2306, which was reverted, to
avoid name conflicts between tasklet connector and data node.

The previous PR was quite invasive, and introduced some field validation
error in ICON4Py. This PR uses a smaller change, very local to the
source of the name conflict: `__arg` is used in GTIR lowering for
expression arguments, which are lowered in the SDFG as data nodes.
Therefore, these names can collide with the tasklet connectors in the
same graph. To avoid this issue, we use a different naming convention
for the tasklet arguments.