fix[next][dace]: SDFG instrumentation with empty else-branch caused memory leak (#2321)
This PR fixes a bug introduced by #2280. The previous PR used a
conditional interstate edge to conditionally execute a sink state, which
contained the timer tasklet to produce the stencil compute time.
However, there was no else-branch, and no unique sink node in the SDFG.
If metrics collection was enabled, the new state would be the sink,
otherwise the predecessor state. The lack of a unique sink node caused
some error in SDFG transformations that led to memory leak, because the
GPU temporary arrays were not properly freed.
This PR builds two conditional blocks, one as unique source and another
one as unique sink node for the SDFG. Inside the conditional block,
metrics collection is represented as a branch with conditional
execution. In this way, the SDFG source and sink nodes are always the
same, no matter whether metrics collection is enabled or not.