test[cartesian]: remove duplicate tests in dace parsing
For some reason, we have these duplicate tests in
`test_dace_parsing.py`. They might have been introduced by a bad merge
or so. For each deleted test there exists another one, which is similar
and - in addition - covers more backends.

This duplication might be related to the test failures that we see on
the daily runs.