test[cartesian]: remove duplicate dace parsing tests (#2325)
## Description

For some reason, we have these duplicate tests in
`test_dace_parsing.py`. They might have been introduced by a bad merge
or so. For each deleted test there exists another one, which is similar
and - in addition - covers more backends.

This duplication might or might not be related for the test failures
that we see on the daily runs. In any case, it's not good practice to
have them duplicated.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>