refactor[next][dace]: cleanup update_sdfg_args (#2312)
This PR is mainly a code refactoring.
The only change is the way the program call decorator iterates on the
arguments: before it accessed the arguments list by index (`args[i]`),
which led to multiple accesses at the same index for field arguments,
because we need to retrieve the field range and the memory strides for
all dimensions. With this change, the arguments list is unpacked first,
then each argument is accessed as a variable (`arg_0`, `arg_1` etc.).
Since this code was on the call path, the code change should bring some
small speedup.