fix: int literal precision for K iterator access
This PR fixes the precision of iterator K access, which was hard coded
to 64bit before. Now, it respects `literal_integer_precision`, which can
be defined per stencil. To do so, a dtype is added at the definition IR
level and then propagated, removing the need to add special treatment
for iterator access in the `gtir_dtype_resolver`.

The PR also changes the access kind of iterator K access at the GTIR
level from `FIELD` (the default for any expression) to `SCALAR`.