fix[dace][next]: Fixed `DeadMapElimination` (#2340)
The case of a Map that had multiple connections to an AccessNode was not
properly handled.
A `list` was used which means that a node could appear twice which lead
to a NetworkX error.
This bug was exposed by the
[`MapToCopy`](https://github.com/GridTools/gt4py/pull/2311)
transformation (this PR must be merged before the `MapToCopy` PR).