feat[dace][next]: Added `MapToCopy` (#2311)
`MapToCopy` is able to transform, under certain conditions, a Map into a
Memlet.
These kind of Maps are created when some slices are accessed in GT4Py.
The transformation is added to `gt_simplify()` but only active in the
initial simplification, since its usefulness is not that big once
`MapFusion*` has started.

---------

Co-authored-by: edopao <edoardo.paone@cscs.ch>