feat[cartesian]: 2D temporaries [Experimental] (#2314)
## Description

Canonically, temporaries have been 3D arrays of float. We introduced a
few months back the capacity to type hint the type to allow for
`integer` and mixed precision temporaries.

Another missing feature is the capacity to ask for a 2D temporary when
the semantic of the code asks for it (aggregate, mask, etc.). We
introduce with two flavors:

- The default behavior knows how to resolve the `FieldDescriptor` type
e.g.
```python
tmp: Field[IJ, np.float64] = ...
```
- A user defined mapping from symbols (as string) to type allows for
applications on top of GT4Py to control the verbosity of the syntax
```python
@stencil(backend=..., dtype={'MyType': Field[IJ, np.float64]})
def stencil(...):
   ...
   tmp: MyType = ...
```

The feature is limited to 2D temporaries for now - and guarded for it.

The feature is experimental because support for `gt:X` backends is
missing.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
- [x] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.