feat[dace][next]: Map Splitter (#2304)
The `SplitMap` transformation is able to split a Map into different
smaller ones, in that sense it is similar to the `VerticalMapSplitRange`
transformation, but targets different cases. The transformation targets
 the following pattern, `MapExit -> (AccessNode)`, in addition `AccessNode`
is only written to by the Map and there is only one consumer.
Furthermore, there is more data written to than is read from the
`AccessNode`, such situation are created by `broadcast()` expression and
`concat_where`s.

Note: This transformation does not integrate the transformation into the
Pipeline, this is instead done in [PR#2284](https://github.com/GridTools/gt4py/pull/2284).