feat[dace][next]: Modified How Degenerated 1D Maps Are Handled (#2349)
The PR modifies how degenerated 1d Maps, i.e. Maps with more then 1
dimension, but only one of them is != 1, are handled.
The previous version also handled changes where the Map had more than 4
dimensions, which leaded to a complicated logic that required to sort
the sizes.
However, in case of not fully specialized Maps, the Maps still might
have symbolic sizes, which made it impossible to sort them, thus this
feature was removed.