build: add script to set the default fallback version to latest release (#2341)
This PR updates the base component of the fallback version information
(used by `version-in-git` when it doesn't find the information anywhere
else) to the latest release (1.0.10), plus explicit segments mentioning
that more concrete information is missing.

For future releases, a new subcommand as been added to the `update`
command of the scripts tool, to set the latest release version as the
base component of the fallback version information. The developers
documentation has been updated accordingly to include this step in the
process to release a new version.