perf[next]: cache low level buffer information in fields (#2342)
Add a new `BufferInfo`/`__gt_buffer_info__` interface to collect the
required low-level information to pass Field buffers to backends and
cache it on the `Field` side, when the `Field` implementation allows it.
The DaCe binding generator already uses this new interface.

Additionally, this PR minor refactorings in the `Field` implementation,
and minor code style changes to make the implementation follow the same
order of method definitions as the interface.

---------

Co-authored-by: Edoardo Paone <edoardo.paone@cscs.ch>
Co-authored-by: Hannes Vogt <hannes@havogt.de>