refactor[next][dace]: Update lowering of domain (#2350)
This PR was created as a side-track of the lowering of multiple output
domains. It refactors the code with the following objectives:
- Cleanup the code to make it possible to apply changes for multiple
output domain.
- Remove `target_domain` from `SubgraphContext`, for two reasons: 1. it
becomes difficult to handle with multiple output domains and 2. it was
applying domain-inference while lowering, which is not needed anymore
after improvements to lowering of `concat_where` expressions (see
#2286).
- Replace dace functions for manipulation of `conat_where` domain with
functions from gt4py `domain_utils` module.

Verification:
- [x] All ICON4Py CI tests pass
- [x] ICON-Blueline stencil verified on `mch_icon-ch1_medium`
- [x] Validation of performance regression in progress