feat[dace][next]: Strides of Views (#2334)
In case there is a view of a global array, we do not have a problem,
because we never modify the strides of globals. The problems start
when there is a view of a transient. The SDFG is constructed using C
order, which good for CPU. However, for GPU we go to FORTRAN order
 which is essentially swapping the memory order. But we can not only
do that for transients, but must also do it for views, which might be
difficult, because views can add or remove certain dimensions.

This PR does _not_ solve it, but instead generates an error if such a
case is detected.