fix[cartesian]: do upcasting on both sides of assignments
In the GTIR layer there's a system to handle operations with different
types. This "GTIR upcaster" was only running on the right hand side of
parallel assignments. However, there are situations where we also have
situations on that left hand side that need upcasting. For example,
users reported an offset calculation with an int32 field and an int64
scalar. This PR fixes the issue by running the GTIR upcaster on both
sides of an assignment.