feat[dace][next] More Roboust Calling (#2353)
This PR changes how calls to the underlying `CompiledSDFG` objects are
carried out. Before, the implementation heavily relied on the internal data
format of the DaCe class. However a recent [change in DaCe](https://github.com/spcl/dace/pull/2185) changed this internal data leading to errors, that had to [be patched](https://github.com/GridTools/dace/pull/9).
This PR introduces a more stable fix and builds upon a [refactoring in DaCe](https://github.com/spcl/dace/pull/2206) that beside other things, exposes the tools that were needed by GT4Py to
work independently of the internals.
For that reason this PR also updates the DaCe dependency to `2025.11.04`.

The main change is, that the argument vector, i.e. the C representation
of the arguments used for the call, are no longer managed by
`CompiledSDFG` but instead by GT4Py's `CompiledDaceProgram`.

Co-authored-by: edopao <edoardo16@gmail.com>