bug[next]: fix timer precision loss (#2364)
Previously we were computing the floating point differences of time
since epoch, i.e. cancellation such that resulting precision was only in
the order of microseconds.

In Python use `time.perf_counter()` instead. In gtfn, we do the
computation with std::chrono::duration, in DaCe we go via nanoseconds
(as int64) since epoch, as we cannot use a std::chrono::time_point in
the interface easily.

---------

Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>