fix[cartesian]: upcast arguments of cast operations
Cast operations are except from upcasting (because they cast
explicitly). With skipping upcasting for cast operations, we also
stopped the recursion, essentially skipping the upcasting of any
argumetns to cast functions. Most of the time, we cast single values so
we didn't run into this until late.

This was reported upstream in NDSL as part of issue
https://github.com/NOAA-GFDL/NDSL/issues/195.