fix[cartesian]: upcast arguments of cast operations (#2369)
## Description

Cast operations are except from upcasting (because they cast
explicitly). With skipping upcasting for cast operations, we also
stopped the recursion, essentially skipping the upcasting of any
arguments to cast functions. Most of the time, we cast single values so
we didn't run into this until late.

This was reported upstream in NDSL as part of issue
https://github.com/NOAA-GFDL/NDSL/issues/195.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>