fix[cartesian]: respect percision of constants (#2377)
## Description

In the `IRMaker`, we discard type information that is potentially
available from inlined constants because e.g. `np.float64` is of
instance `float`, which means, that the old order of the `if` statements
would trigger for `isinstance(..., (int, float))` in cases that we don't
want. We flip the order to ensure that we keep potential type
annotations on e.g. global constants.

This is fallout from literal precision work (see PR #2187) and was
reported (first hidden by other issues) in upstream issue
https://github.com/NOAA-GFDL/NDSL/issues/195.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>