fix[next][dace] Set properly strides of `data` in CPU (#2384)
After https://github.com/GridTools/gt4py/pull/2355 we have to update the strides of AccessNode `data` for memlets inside `NestedSDFG`s properly.
`gt_change_strides` used to run only for the GPU backend. Now it's necessary for CPU as well.
The final strides must be set after all the transformations are done because some AccessNodes might be removed and the strides of the corresponding AccessNodes inside `NestedSDFG`s need to be updated

---------

Co-authored-by: Philip Müller <147368808+philip-paul-mueller@users.noreply.github.com>