bug[next]: disallow implicit bool conversion of fields (#2393)
An old NumPy version caught the following mistake `np.cumsum(field)`,
which should have been `np.cumsum(field.asnumpy())`. The result was a
size-1 NumPy object-array which contained a Field. The mistake was not
caught because `bool(field)==True`, `np.allclose` broadcasted the size-1
array to the size of the other array and did math operations that where
all supported by the `Field` API which in the end resulted in an np.all
check of an array of Fields, which is always `True`.

Long story short, we shouldn't allow implicit conversion of Fields to
bool, which revealed more bugs in our tests.

---------

Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>