refactor[next][dace]: Simplify visitor of `concat_where` expressions (#2394)
This PR reduces the number of lines of code in the lowering of
`concat_where` to SDFG. This is achieved mainly by means of two changes:
- Reuse the logic for broadcasting of a scalar and of a field slice.
Both kinds of input can be handled as a broadcast expression, which is
currently lowered as a mapped tasklet. In a separate PR, we will
introduce a broadcast library node for this.
- Handle the lower and upper domain of the concat domain with the same
utility function (`_translate_concat_where_branch()`) since the
difference between these two cases is minimal. The function argument
`is_lower` allows to select one or the other.