fix[next][dace]: Use explicit dataflow in let-lambdas (#2396)
This PR was created in the context of investigating the result
validation failure in the `graupel` program. Note that this PR does not
solve the bug, but it provides a better SDFG representation of
let-lambdas.
The domain inference pass can detect that certain arguments are not
used, and do not need to be computed. The corresponding access nodes are
not created, in the SDFG, and we use `None` in the construction of the
intermediate result for the lowering. However, on the interface to
nested SDFGs, which are used to lower let-lambda nodes, the `None`
arguments were not properly handled and resulted in transient nodes with
uninitialized data. DaCe would just emit a warning, during validation,
but would still accept this SDFG. This PR ensures that `None` nodes do
not end up in the SDFG, and all arguments passed to a nested SDFG
contain defined values, either globals or temporary nodes with the
result of some intermediate computation.

- [x] ICON4Py CI tests pass
- [x] Validated in ICON-blueline
- [x] No performance regression in ICON-blueline