feat[cartesian]: Runtime Interval Bounds (#2395)
## Description
In the context of porting physics parametrization, a repeating pattern
in a lot of solver-like codes are column based loops that only affect
parts of the column until a certain condition is met.
This PR implements the feature of allowing specification of the interval
at runtime as an experimental feature - meaning we only have an
implementation in the debug backend.
For more details, consut the corresponding ADR.

## Requirements

- [ ] All fixes and/or new features come with corresponding tests.
- [x] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.

---------

Co-authored-by: Roman Cattaneo <romanc@users.noreply.github.com>