fix[next][dace]: Use symbol mapping for stride propagation across nested SDFGs (#2404)
This PR addresses a validation error observed in the `graupel` program
in ICON4Py. The validation of this program failed on the `t` output
field, although the computation of `t` was correct. The write to another
output field, `pr`, which was not verified in the `graupel` test, was
writing out of buffer because of wrong strides. The strides were
incorrectly set during SDFG transformations because the symbol mapping
was wrong for a specific lowering case, where the same field name `pr`
was used inside the scope of a scan field operator and in the parent
scope of a let-lambda expression which was accessing the `pr` program
argument. This PR fixes the symbol mapping in strides propagation.